requirements.txt
google-api-python-client>=2.166.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
pdfplumber>=0.11.5
pypdf>=5.4.0
python-dotenv>=1.0.1
pytesseract>=0.3.13
Pillow>=11.0.0
pdf2image==1.17.0
opencv-python>=4.10.0.84
numpy>=2.1.0
google-genai>=1.68.0